update to chromium v60.0.3112.20, fixes build problem.

+++ app/src/main/java/org/chromium/chrome/browser/compositor/layouts/content/ContentOffsetProvider.java
@@ -10,7 +10,8 @@ package org.chromium.chrome.browser.compositor.layouts.content;
  */
 public interface ContentOffsetProvider {
     /**
-     * @return How far to translate any Android overlay views to recreate the correct content data.
+     * @return How far to translate any Android overlay views to recreate the correct content data
+     *         in px.
      */
-    public int getOverlayTranslateY();
+    public float getOverlayTranslateY();
 }
